Add test for removing deleted image from mark list
The test currently fails but should not fail anymore once the fix is
implemented.

It required some restructuring of the bdd steps as it makes use of steps
that were local to other parts.
